+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Next
 ===
 
+- new command `jcommunity:user:import` to import accounts from a CSV file
+
+
 Version 1.4.5
 =============